pyproject.toml
[project]
name = "genevariantfetcher"
version = "0.1.0"
description = "A collection of Python tools for extracting genetic variant data, patient-level information, and full-text articles from PubMed literature"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "biopython>=1.86",
    "litellm>=1.30.0",
    "markitdown>=0.1.3",
    "openai>=2.8.1",
    "pandas>=2.3.3",
    "pydantic>=2.0.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.5",
    "spyprot>=0.9.6",
    "tabulate>=0.9.0",
    "tenacity>=9.1.2",
    "typer>=0.9.0",
]

[project.scripts]
gvf = "pipeline.cli:app"
